Moved code for accessing oura API
