+++ lib/widgets/OuraLoginButton.dart
@@ -1,13 +1,7 @@
-import 'dart:html';
-import 'dart:io';
-// ignore_for_file: depend_on_referenced_packages
-import 'package:http/http.dart' as http;
 // ignore: avoid_web_libraries_in_flutter
 import 'dart:html' as html;
-import 'dart:convert';
 import 'package:flutter/material.dart';
 import 'package:ouraintervention/misc/Database.dart';
-import 'package:ouraintervention/objects/SleepData.dart';
 
 const clientId = 'Q4EDXKDK2244IHFU';
 const clientSecret = 'VYOANLDLNJ473P6W7YAA2FRB6C4KRBXX';
@@ -66,19 +60,8 @@ class _OuraLoginButtonState extends State<OuraLoginButton> {
           _loginButtonText = 'Oura Ring Connected';
         });
         var token = await _login(event.data, popupWindow);
-        var url = Uri.parse('https://api.ouraring.com/v1/sleep?start=2020-01-01&end=2022-04-22&access_token=$token');
-        var response = await http.get(url, headers: {"Accept": "application/json", "Access-Control-Allow-Origin": "*"});
+        await widget.database.uploadSleepData(token);
 
-        // Decode the string from HTTP request to API
-        Map<String, dynamic> sleepJson = jsonDecode(response.body) as Map<String, dynamic>;
-
-        // Translate it so that Database method can be used to upload
-        List<SleepData> sleepList = [];
-        for (int i = 0; i < sleepJson['sleep']!.length; i++) {
-          sleepList.add(SleepData.fromJson(sleepJson['sleep']![i]));
-        }
-
-        widget.database.uploadSleepData(sleepList);
         widget.database.getActionDates('run');
       }
     });
